onsik-mvp/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
supabase>=2.3.0
celery>=5.3.0
redis>=5.0.0
scrapy>=2.11.0
beautifulsoup4>=4.12.0
langchain>=0.3.0
langchain-huggingface>=0.1.0
sentence-transformers>=2.0.0
transformers>=4.43.0
torch
accelerate
bitsandbytes
requests>=2.31.0
python-dotenv>=1.0.0
langchain-openai>=0.2.0
